+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @verdaccio/cli
 
+## 8.0.0-next-8.25
+
+### Patch Changes
+
+- 3aff890: chore(deps): api, auth, cli, config, url
+- Updated dependencies [3aff890]
+- Updated dependencies [b24f513]
+- Updated dependencies [6d1a84a]
+  - @verdaccio/config@8.0.0-next-8.25
+  - @verdaccio/node-api@8.0.0-next-8.25
+  - @verdaccio/core@8.0.0-next-8.25
+  - @verdaccio/logger@8.0.0-next-8.25
+
 ## 8.0.0-next-8.24
 
 ### Patch Changes