+++ packages/tools/generator-verdaccio-plugin/CHANGELOG.md
@@ -1,5 +1,14 @@
 # generator-verdaccio-plugin
 
+## 6.0.0-next-8.25
+
+### Patch Changes
+
+- Updated dependencies [3aff890]
+- Updated dependencies [b24f513]
+  - @verdaccio/config@8.0.0-next-8.25
+  - @verdaccio/core@8.0.0-next-8.25
+
 ## 6.0.0-next-8.24
 
 ### Patch Changes